Final build Removed unused files from project Added missing assets to Git project Fixed module upgrade bug (array out of bounds) Cleaned up code (removed warnings, todo tags, and typos) Changed upgrade button text when maximum level if reached

+++ core/src/com/nic/projectevolve/screens/MenuScreen.java
@@ -42,7 +42,7 @@ public class MenuScreen implements Screen {
     public MenuScreen(ProjectEvolve game) {
         create();
         this.game = game;
-        ProjectEvolve.setScl((int) (Gdx.graphics.getHeight() / 8));
+        ProjectEvolve.setScl(Gdx.graphics.getHeight() / 8);
 
         //OrthographicCamera gameCam;
         gameCam = new OrthographicCamera();
